+++ lib/pages/main_page/games_widget_option.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class GamesWidgetOption extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Text('Games');
+  }
+}