+++ lib/pages/main_page/search_widget_option.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class SearchWidgetOption extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Text('Search');
+  }
+}