+++ lib/pages/main_page/statistics_widget_option.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class StatisticsWidgetOption extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Text('Statistics');
+  }
+}